remove PageProcessorHelper
