+++ packages/crawler/src/page-processors/classic-page-processor.spec.ts
@@ -8,16 +8,16 @@ import { IMock, Mock } from 'typemoq';
 import { AccessibilityScanOperation } from '../page-operations/accessibility-scan-operation';
 import { BlobStore, DataStore } from '../storage/store-types';
 import { ClassicPageProcessor } from './classic-page-processor';
-import { PageProcessorHelper } from './page-processor-helper';
+import { PartialScanData } from './page-processor-base';
 
 // tslint:disable: no-any
 
 describe(ClassicPageProcessor, () => {
-    let helperMock: IMock<PageProcessorHelper>;
     let requestQueueMock: IMock<Apify.RequestQueue>;
     let accessibilityScanOpMock: IMock<AccessibilityScanOperation>;
     let dataStoreMock: IMock<DataStore>;
     let blobStoreMock: IMock<BlobStore>;
+    let enqueueLinksExtMock: IMock<typeof Apify.utils.enqueueLinks>;
 
     const testUrl = 'test url';
     const testId = 'test id';
@@ -28,11 +28,11 @@ describe(ClassicPageProcessor, () => {
     let classicPageProcessor: ClassicPageProcessor;
 
     beforeEach(() => {
-        helperMock = Mock.ofType<PageProcessorHelper>();
         requestQueueMock = Mock.ofType<Apify.RequestQueue>();
         accessibilityScanOpMock = Mock.ofType<AccessibilityScanOperation>();
         dataStoreMock = Mock.ofType<DataStore>();
         blobStoreMock = Mock.ofType<BlobStore>();
+        enqueueLinksExtMock = Mock.ofType<typeof Apify.utils.enqueueLinks>();
         requestStub = {
             id: testId,
             url: testUrl,
@@ -45,30 +45,49 @@ describe(ClassicPageProcessor, () => {
 
         classicPageProcessor = new ClassicPageProcessor(
             requestQueueMock.object,
-            helperMock.object,
             discoveryPatterns,
             accessibilityScanOpMock.object,
             dataStoreMock.object,
             blobStoreMock.object,
+            enqueueLinksExtMock.object,
         );
     });
 
     afterEach(() => {
-        helperMock.verifyAll();
+        enqueueLinksExtMock.verifyAll();
         accessibilityScanOpMock.verifyAll();
+        blobStoreMock.verifyAll();
     });
 
     it('pageProcessor', async () => {
-        helperMock.setup((h) => h.enqueueLinks(pageStub, requestQueueMock.object, discoveryPatterns)).verifiable();
+        setupEnqueueLinks(pageStub);
         accessibilityScanOpMock.setup((aso) => aso.run(pageStub, testId, blobStoreMock.object)).verifiable();
         const expectedScanData = {
             id: testId,
             url: testUrl,
+            succeeded: true,
         };
-        helperMock.setup((h) => h.pushScanData(blobStoreMock.object, expectedScanData)).verifiable();
+        setupPushScanData(expectedScanData);
 
         // tslint:disable-next-line: no-any
         const inputs: Apify.PuppeteerHandlePageInputs = { page: pageStub, request: requestStub } as any;
         await classicPageProcessor.pageProcessor(inputs);
     });
+
+    function setupEnqueueLinks(page: Page): void {
+        const options = {
+            page: pageStub,
+            requestQueue: requestQueueMock.object,
+            pseudoUrls: discoveryPatterns,
+        };
+        enqueueLinksExtMock
+            .setup((el) => el(options))
+            .returns(async () => [])
+            .verifiable();
+    }
+
+    function setupPushScanData(expectedScanData: PartialScanData): void {
+        const id = `${expectedScanData.id}.data`;
+        blobStoreMock.setup((bs) => bs.setValue(id, expectedScanData)).verifiable();
+    }
 });
