+++ packages/crawler/src/page-processors/classic-page-processor.ts
@@ -6,8 +6,8 @@ import { PageProcessorBase } from './page-processor-base';
 export class ClassicPageProcessor extends PageProcessorBase {
     public pageProcessor: Apify.PuppeteerHandlePage = async ({ page, request }) => {
         console.log(`Crawling page ${page.url()}`);
-        await this.helper.enqueueLinks(page, this.requestQueue, this.discoveryPatterns);
+        await this.enqueueLinks(page);
         await this.accessibilityScanOp.run(page, request.id as string, this.blobStore);
-        await this.helper.pushScanData(this.blobStore, { id: request.id as string, url: request.url });
+        await this.pushScanData({ id: request.id as string, url: request.url });
     };
 }
