+++ packages/crawler/src/page-processors/page-processor-base.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import Apify from 'apify';
+import { Page } from 'puppeteer';
 import { AccessibilityScanOperation } from '../page-operations/accessibility-scan-operation';
 import { ScanData } from '../scan-data';
 import { LocalBlobStore } from '../storage/local-blob-store';
 import { LocalDataStore } from '../storage/local-data-store';
 import { BlobStore, DataStore, scanResultStorageName } from '../storage/store-types';
-import { PageProcessorHelper } from './page-processor-helper';
 
 export interface PageProcessorOptions {
     baseUrl: string;
@@ -15,6 +15,11 @@ export interface PageProcessorOptions {
     simulate?: boolean;
     selectors?: string[];
 }
+
+export type PartialScanData = {
+    url: string;
+    id: string;
+} & Partial<ScanData>;
 
 export interface PageProcessor {
     pageProcessor: Apify.PuppeteerHandlePage;
@@ -37,11 +42,11 @@ export abstract class PageProcessorBase implements PageProcessor {
 
     public constructor(
         protected readonly requestQueue: Apify.RequestQueue,
-        protected readonly helper: PageProcessorHelper,
         protected readonly discoveryPatterns?: string[],
         protected readonly accessibilityScanOp: AccessibilityScanOperation = new AccessibilityScanOperation(),
         protected readonly dataStore: DataStore = new LocalDataStore(scanResultStorageName),
         protected readonly blobStore: BlobStore = new LocalBlobStore(scanResultStorageName),
+        private readonly enqueueLinksExt: typeof Apify.utils.enqueueLinks = Apify.utils.enqueueLinks,
         private readonly gotoExtended: typeof Apify.utils.puppeteer.gotoExtended = Apify.utils.puppeteer.gotoExtended,
     ) {}
 
@@ -70,4 +75,31 @@ export abstract class PageProcessorBase implements PageProcessor {
         };
         await this.dataStore.pushData(scanData);
     };
+
+    protected async enqueueLinks(page: Page): Promise<Apify.QueueOperationInfo[]> {
+        const enqueued = await this.enqueueLinksExt({
+            page,
+            requestQueue: this.requestQueue,
+            pseudoUrls: this.discoveryPatterns,
+        });
+        console.log(`Discovered ${enqueued.length} links on page ${page.url()}`);
+
+        return enqueued;
+    }
+
+    protected async pushScanData(scanData: PartialScanData): Promise<void> {
+        const mergedScanData: ScanData = {
+            succeeded: true,
+            ...scanData,
+        };
+        await this.blobStore.setValue(`${scanData.id}.data`, mergedScanData);
+    }
+
+    // protected async saveSnapshot(page: Page, id: string): Promise<void> {
+    //     await Apify.utils.puppeteer.saveSnapshot(page, {
+    //         key: `${id}.screenshot`,
+    //         saveHtml: false,
+    //         keyValueStoreName: scanResultStorageName,
+    //     });
+    // }
 }
